refactor(node): separate embedding from runtime bootstrap (#947)
* refactor(node): separate embedding from runtime bootstrap

* fix(node): clean up embedding test diagnostics

* fix(node): include runtime chunk in Sentry source maps
