+++ .changeset/node-embedding.md
@@ -1,0 +1,5 @@
+---
+"hevy-mcp": patch
+---
+
+Make the package-root Node API safe for embedding by separating it from executable telemetry and lifecycle startup.
