+++ package.json
@@ -47,7 +47,7 @@
 		"test:integration": "vitest run tests/integration",
 		"test:diagnostics": "node --test tests/nightly/diagnostics.test.mjs",
 		"test:contract": "vitest run packages/node/src/index.test.ts packages/node/src/utils/config.test.ts packages/core/src/tools/register.test.ts packages/core/src/utils/output-schemas.test.ts tests/unit/server-manifest.test.ts",
-		"test:stdio": "vitest run packages/node/src/index.test.ts packages/node/src/utils/stdio-observability.test.ts packages/node/src/utils/graceful-shutdown.test.ts packages/node/src/utils/graceful-shutdown.child-process.test.ts",
+		"test:stdio": "vitest run packages/node/src/index.test.ts packages/node/src/runtime.test.ts packages/node/src/utils/stdio-observability.test.ts packages/node/src/utils/graceful-shutdown.test.ts packages/node/src/utils/graceful-shutdown.child-process.test.ts",
 		"test:worker-http": "vitest run tests/integration/worker-http.integration.test.ts",
 		"test:worker": "vitest run --config vitest.workers.config.ts",
 		"test:worker-http:live": "node --env-file-if-exists=.env scripts/run-live-vitest.mjs HEVY_RUN_LIVE_WORKER_TESTS=1,HEVY_API_KEY tests/integration/worker-http.live.integration.test.ts",
