+++ packages/node/tsdown.config.ts
@@ -55,6 +55,59 @@ if (
 		)}, version=${String(version)}`,
 	);
 }
+
+const packageName = name as string;
+const packageVersion = version as string;
+
+function createReleaseSentryPlugin() {
+	const [plugin] = sentryRollupPlugin({
+		org: process.env.SENTRY_ORG,
+		project: process.env.SENTRY_PROJECT,
+		authToken: process.env.SENTRY_AUTH_TOKEN,
+		telemetry: false,
+		sourcemaps: {
+			assets: ["./dist/**/*.mjs", "./dist/**/*.map"],
+			filesToDeleteAfterUpload: ["./dist/**/*.map"],
+		},
+		release: {
+			name: `${packageName}@${packageVersion}`,
+			inject: false,
+		},
+	});
+	if (!plugin) throw new Error("Sentry release plugin was not created");
+
+	return {
+		...plugin,
+		renderChunk(
+			code: string,
+			chunk: { facadeModuleId?: string | null; moduleIds?: string[] },
+			options: unknown,
+			meta: unknown,
+		) {
+			const normalizeModuleId = (moduleId: string) =>
+				moduleId.replaceAll("\\", "/");
+			const isExecutableModule = (moduleId: string | null | undefined) => {
+				const normalized = moduleId && normalizeModuleId(moduleId);
+				return (
+					normalized === "src/cli.ts" ||
+					normalized?.endsWith("/src/cli.ts") ||
+					normalized === "src/runtime.ts" ||
+					normalized?.endsWith("/src/runtime.ts")
+				);
+			};
+			// The published package root is an embedding surface. Inject Sentry
+			// debug IDs into the executable entry and its lazy runtime chunk;
+			// shared chunks imported by the embedding entry must remain inert.
+			const isExecutableChunk = [
+				chunk.facadeModuleId,
+				...(chunk.moduleIds ?? []),
+			].some(isExecutableModule);
+			if (!isExecutableChunk) return null;
+			return plugin.renderChunk?.(code, chunk, options, meta);
+		},
+	};
+}
+
 export default defineConfig({
 	entry: isStandaloneBuild ? ["src/cli.ts"] : ["src/index.ts", "src/cli.ts"],
 	format: ["esm"],
@@ -111,21 +164,7 @@ export default defineConfig({
 		},
 	},
 	plugins: [
-		sentryRollupPlugin({
-			org: process.env.SENTRY_ORG,
-			project: process.env.SENTRY_PROJECT,
-			authToken: process.env.SENTRY_AUTH_TOKEN,
-			telemetry: false,
-			sourcemaps: {
-				assets: ["./dist/**/*.mjs", "./dist/**/*.map"],
-				...(isReleaseBuild
-					? { filesToDeleteAfterUpload: ["./dist/**/*.map"] }
-					: {}),
-			},
-			release: {
-				name: `${name}@${version}`,
-			},
-		}),
+		...(isReleaseBuild ? [createReleaseSentryPlugin()] : []),
 		...(enableCodecovBundleAnalysis
 			? codecovRollupPlugin({
 					enableBundleAnalysis: true,
