+++ repository/validation-lanes.json
@@ -88,6 +88,7 @@
 				"kind": "vitest",
 				"include": [
 					"packages/node/src/index.test.ts",
+					"packages/node/src/runtime.test.ts",
 					"packages/node/src/utils/stdio-observability.test.ts",
 					"packages/node/src/utils/graceful-shutdown.test.ts",
 					"packages/node/src/utils/graceful-shutdown.child-process.test.ts"
